+++ sshd-core/src/test/java/org/apache/sshd/client/channel/ChannelExecTest.java
@@ -88,7 +88,9 @@ public class ChannelExecTest extends BaseTestSupport {
 
     @Test   // see SSHD-692
     public void testMultipleRemoteCommandExecutions() throws Exception {
-        try (ClientSession session = client.connect(getCurrentTestName(), TEST_LOCALHOST, port).verify(7L, TimeUnit.SECONDS).getSession()) {
+        try (ClientSession session = client.connect(getCurrentTestName(), TEST_LOCALHOST, port)
+                .verify(7L, TimeUnit.SECONDS)
+                .getSession()) {
             session.addPasswordIdentity(getCurrentTestName());
             session.auth().verify(5L, TimeUnit.SECONDS);